Implemented TrendingScreen.kt and necessary functionality in TitlesRemoteDataSource.kt and TitlesRepositoryImpl.kt and TmdbApi.kt. Renamed SearchApiResponse into TitlesListResponse, because same class is used for trending as well because same structure of data is returned from the api.
Hoisted up getting of the title item poster to the Main Activity and providing it as a dependency to all screens.

Removed "TODO: TEMP IMPLEMENTATION FOR TESTING" comments from SearchViewModel.kt. This functionality will stay like this for now. In the future this might get changed if we start using mapped/filtered flows.