WIP: Videos in DetailsScreen.kt. Remade MyYoutubeLinkExtractor.kt and the library's YouTubeExtractor.java to use THREAD_POOL_EXECUTOR so that AsyncTasks are being run in parallel.
Remade the DetailsViewModel, DetailsUiState.kt and DetailsScreen.kt to use Loading, Ready and Error states of DetailsUiState.

WIP: continuing to work on providing the youtube links for the video player. Currently a filtered YtVideo list is available that filters for formats we want (defined in Constants.kt). Next steps: get more video data from Api, create a new entity and and set up relationship between it and Title and get thumbnails from Yt.