WIP: Pagination for Trending Titles. Added Paging dependencies. Created TitleItemCache file that holds all entities for cached TitleItems. So far just for Trending. Created TitleItemCachedMapper.kt to map TitleItemFull to required cache database entities. Created TitlesTrendingRemoteMediator.kt for handling mediation of data from api to local cache.
Added temporary-named methods of getTrendingTitlesPaging to repository and remote data source.

Added a typeCode value to TitleType and TitleTypeApiModel in order to have actual usable ids for TitleItemApiModel and TitleItemFull that are put together from the code (TV/MOVIE) and the mediaId. Adjusted MyGsonConverter.kt accordingly to assign this value already there. Subsequent mappers will take that id.

Created CacheDao interface to hold methods for handling the cache. Added it to AppModule for DI.